requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
matplotlib
joblib
transformers
groq
python-dotenv
